app/config.py
import os

from dotenv import load_dotenv

load_dotenv()

class Settings:
    DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./test.db")
    SECRET_KEY=os.getenv("SECRET_KEY", "secret key")

settings = Settings()
